feat(page-service): sync routes with kitraw.json on page creation and deletion

+++ src/services/page-service.ts
@@ -9,13 +9,24 @@ export class PageService {
     const group: string = isPrivate ? "private" : "public";
     const pagePath: string = path.join(process.cwd(), "src", "app", "[locale]", `(${group})`, name);
     const groupLayoutPath: string = path.join(process.cwd(), "src", "app", "[locale]", `(${group})`, "layout.tsx");
-
     const folderExists: boolean = await this.fileSystem.exists(pagePath);
+    const kitrawPath: string = path.join(process.cwd(), "kitraw.json");
+    const kitrawExists: boolean = await this.fileSystem.exists(kitrawPath);
+    const kitrawConfig: any = await this.fileSystem.readJson(kitrawPath);
+    
+    if (!kitrawExists) {
+      throw new Error(`File ${kitrawPath} does not exist`);
+    }
 
     if (folderExists) {
       throw new Error(`Folder ${name} already exists`);
     } else {
       await this.fileSystem.writeFile(path.join(pagePath, "page.tsx"), this.pageTemplate(name));
+
+      kitrawConfig["routes"][group + "Routes"].push(name);
+
+      await this.fileSystem.writeJson(kitrawPath, kitrawConfig);
+
       const groupLayoutExists: boolean = await this.fileSystem.exists(groupLayoutPath);
       if (!groupLayoutExists) {
         if (group === "public") {
@@ -31,9 +42,22 @@ export class PageService {
     const group: string = isPrivate ? "private" : "public";
     const pagePath: string = path.join(process.cwd(), "src", "app", "[locale]", `(${group})`, name);
     const folderExists: boolean = await this.fileSystem.exists(pagePath);
+    const kitrawPath: string = path.join(process.cwd(), "kitraw.json");
+    const kitrawExists: boolean = await this.fileSystem.exists(kitrawPath);
+    const kitrawConfig: any = await this.fileSystem.readJson(kitrawPath);
+    
+    if (!kitrawExists) {
+      throw new Error(`File ${kitrawPath} does not exist`);
+    }
 
     if (folderExists) {
       await this.fileSystem.remove(pagePath);
+
+      const routes: Array<string> = kitrawConfig["routes"][group + "Routes"];
+      const newRoutes: Array<string> = routes.filter((route: string) => route !== name);
+      kitrawConfig["routes"][group + "Routes"] = newRoutes;
+
+      await this.fileSystem.writeJson(kitrawPath, kitrawConfig);
     } else {
       throw new Error(`Folder ${name} does not exist`);
     }
